requirements.txt
pytesseract
opencv-python
matplotlib
Pillow
PyQt5
jsonpickle
pyinstaller
easyocr